Create demultiplexPairedSamples.pl
The data is generated using a nested multiplex PCR strategy. This script takes the sam output from a search of the multiplex R2 reads (that should be internal to transposable elements) against the TEs and uses that information to demultiplex the R1 and R2 reads based on the TE element (if any) that was hit. The output is put into a file with a given basename. The output is a count of the number of reads associated with each TE element.